+++ fakeredis/_commands.py
@@ -369,3 +369,15 @@ def delete_keys(*keys):
             done.add(key.key)
             ans += 1
     return ans
+
+
+def fix_range(start, end, length):
+    # Redis handles negative slightly differently for zrange
+    if start < 0:
+        start = max(0, start + length)
+    if end < 0:
+        end += length
+    if start > end or start >= length:
+        return -1, -1
+    end = min(end, length - 1)
+    return start, end + 1
